# src/routes/task.routes.js
import express from 'express';
import { auth } from '../middleware/auth.js';
import { taskValidation } from '../validators/task.validator.js';
import {
  getAllTasks,
  getTaskById,
  createTask,
  updateTaskById,
  deleteTaskById
} from '../controllers/task.controller.js';

export const taskRoutes = express.Router();

/**
 * @swagger
 * /api/tasks:
 *   get:
 *     security:
 *       - bearerAuth: []
 *     summary: Get all tasks
 *     tags: [Tasks]
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *           default: 1
 *         description: Page number
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *           default: 10
 *         description: Number of items per page
 *     responses:
 *       200:
 *         description: List of tasks
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 tasks:
 *                   type: array
 *                   items:
 *                     $ref: '#/components/schemas/Task'
 *                 pagination:
 *                   type: object
 *                   properties:
 *                     total:
 *                       type: integer
 *                     page:
 *                       type: integer
 *                     pages:
 *                       type: integer
 *       401:
 *         description: Unauthorized
 */
taskRoutes.get('/', auth, getAllTasks);

/**
 * @swagger
 * /api/tasks/{id}:
 *   get:
 *     security:
 *       - bearerAuth: []
 *     summary: Get task by ID
 *     tags: [Tasks]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *         description: Task ID
 *     responses:
 *       200:
 *         description: Task details
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Task'
 *       404:
 *         description: Task not found
 *       401:
 *         description: Unauthorized
 */
taskRoutes.get('/:id', auth, getTaskById);

/**
 * @swagger
 * /api/tasks:
 *   post:
 *     security:
 *       - bearerAuth: []
 *     summary: Create a new task
 *     tags: [Tasks]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - title
 *             properties:
 *               title:
 *                 type: string
 *               description:
 *                 type: string
 *               status:
 *                 type: string
 *                 enum: [pending, in_progress, completed]
 *                 default: pending
 *     responses:
 *       201:
 *         description: Task created successfully
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Task'
 *       400:
 *         description: Invalid input
 *       401:
 *         description: Unauthorized
 */
taskRoutes.post('/', auth, taskValidation, createTask);

/**
 * @swagger
 * /api/tasks/{id}:
 *   put:
 *     security:
 *       - bearerAuth: []
 *     summary: Update task by ID
 *     tags: [Tasks]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *         description: Task ID
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               title:
 *                 type: string
 *               description:
 *                 type: string
 *               status:
 *                 type: string
 *                 enum: [pending, in_progress, completed]
 *     responses:
 *       200:
 *         description: Task updated successfully
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Task'
 *       404:
 *         description: Task not found
 *       401:
 *         description: Unauthorized
 */
taskRoutes.put('/:id', auth, taskValidation, updateTaskById);

/**
 * @swagger
 * /api/tasks/{id}:
 *   delete:
 *     security:
 *       - bearerAuth: []
 *     summary: Delete task by ID
 *     tags: [Tasks]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *         description: Task ID
 *     responses:
 *       204:
 *         description: Task deleted successfully
 *       404:
 *         description: Task not found
 *       401:
 *         description: Unauthorized
 */
taskRoutes.delete('/:id', auth, deleteTaskById);